Move lower_params/lib to lower_params/lower
